back/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
Pillow
boto3
django-storages
celery
redis
